requirements.txt
groq
SpeechRecognition
pyttsx3
streamlit
python-dotenv